+++ from_ground_up.js
@@ -5,13 +5,16 @@
 	 * @author jacob
 	 * @version 0.1
 	 */
-	 // jsdoc comments are extra indented because i use indent folding in vi
-	 // and i don't want to see the documentation without explicitly unfolding it
 
-	 // after i was told not to use classes, i did some research on the proper,
-	 // `javascript' way of doing things; and was told that the gods of yore
-	 // never used this and rarely used new. I think the only code that uses this
-	 // is Query, because i didn't care enough to fully port it.
+	/*
+	 * jsdoc comments are extra indented because i use indent folding in vi
+	 * and i don't want to see the documentation without explicitly unfolding it
+
+	 * after i was told not to use classes, i did some research on the proper,
+	 * `javascript' way of doing things; and was told that the gods of yore
+	 * never used this and rarely used new. I think the only code that uses this
+	 * is Query, because i didn't care enough to fully port it.
+	 */
 
 const {Readable} = require('stream');
 const xml = require('xml');
@@ -20,7 +23,8 @@ const {Client} = require('pg');
 const archiver = require('archiver');
 // let geojsonhint = require('geojsonhint');
 
-	/** object mapping from languages to database tables to the names presented to the user */
+	/** object mapping from languages to database tables to the names presented to the user 
+	 */
 const names = {
 	en: {
 		areas_vw: 'Area',
@@ -263,57 +267,6 @@ function promise_of_geojson(area_id, client, queries, Feature) {
 		});
 	});
 }
-
-/*
-function upload_to_mapbox(geojson_doc, area_id) {
-	const MY_ACCESS_TOKEN = 'testest'; //TODO
-	const USERNAME = 'testest'; //TODO
-	const AWS = require('aws-sdk'); //TODO move to top
-	const mbxUploads = require('@mapbox/mapbox-sdk/services/uploads');
-	const uploadsClient = mbxClient({accessToken: MY_ACCESS_TOKEN});
-	const region = 'us-east-1';
-	const geojson_stream = new Readable();
-	geojson_stream.push(geojson_doc);
-	geojson_stream.push(null);
-
-	const getCredentials = () => {
-		return uploadsClient
-			.createUploadCredentials()
-			.send()
-			.then(response => response.body);
-	};
-
-	const putFileOnS3 = (credentials) => {
-		const s3 = new AWS.S3({
-			accessKeyId = credentials.accessKeyId,
-			secretAccessKey = credentials.secretAccessKey,
-			sessionToken = credentials.sessionToken,
-			region: REGION
-		});
-		return s3.putObject({
-			Bucket: credentials.bucket,
-			Key: credentials.key,
-			//Body: fs.createReadStream(GEOJSON_FILE_PATH)
-			Body: geojson_stream
-		}).promise();
-	};
-
-	getCredentials()
-		.then(creds => {
-			putFileOnS3(creds)
-				.then(dunno => {
-					uploadsClient.createUpload({
-						mapId: `${USERNAME}.${area_id}`, //TODO
-						url: creds.url
-					})
-						.send()
-						.then(response => {
-							const upload = response.body;
-						});
-				});
-		}); //test for failure
-}
-*/
 
 	/**
 	 * prints geojson to console
